Making all the grades float64.

+++ args/args.go
@@ -3,5 +3,5 @@ package args
 type Args struct {
 	Name    string
 	Subject string
-	Grade   uint
+	Grade   float64
 }
